update: Prefecture Context| stateデータ作成部分を別ファイルに分離
他のテスト時、モックデータから処理作成しやすくするため分離
